# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.1.1",
    "title": "Backend API",
    "description": "Backend App Mobile for LOT Internacional",
    "contact": {
      "name": "marco.bardales@lotinternacional.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost/",
      "description": "TEST SERVER"
    },
    {
      "url": "https://backend-app-lot.onrender.com/",
      "description": "PRODUCTION SERVER (free)"
    }
  ],
  "paths": {
...
    "/docs/swagger.json": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api.services/all": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "Endpoint para obtener una lista de los servicios.",
        "description": "Endpoint para obtener los servicios que ofrece LOT Internacional.",
        "operationId": "api.services/all",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/AllServicesResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/AllServicesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/BadRequestError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/BadRequestError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/api.services/new": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Endpoint para crear un nuevo servicio.",
        "description": "Endpoint para crear un nuevo servicio.",
        "operationId": "api.services/new",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NonContentResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/NonContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/BadRequestError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/BadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/UnauthorizedError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ConflictError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/ConflictError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "required": true,
                    "type": "string"
                  },
                  "thumbnail": {
                    "required": true,
                    "type": "string",
                    "format": "binary"
                  },
                  "pdf": {
                    "required": true,
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api.services/{serviceId}": {
      "delete": {
        "tags": [
          "Services"
        ],
        "summary": "Endpoint para eliminar un servicio.",
        "description": "Endpoint para eliminar un servicio.",
        "operationId": "api.services/:serviceId",
        "parameters": [
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Id de servicio"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NonContentResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/NonContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/BadRequestError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/BadRequestError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/UnauthorizedError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/UnauthorizedError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/NotFoundError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/NotFoundError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ConflictError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/ConflictError"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/responses/InternalServerError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
...
      "AllServicesResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status",
            "example": "success"
          },
          "data": {
            "type": "array",
            "description": "Data",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "description": "Service id",
                  "example": "..."
                },
                "title": {
                  "type": "string",
                  "description": "Title",
                  "example": "..."
                },
                "thumbnailUrl": {
                  "type": "string",
                  "description": "Thumbnail url",
                  "example": "..."
                },
                "pdfUrl": {
                  "type": "string",
                  "description": "Pdf url",
                  "example": "..."
                }
              }
            }
          }
        }
      },
...
      "NonContentResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status",
            "example": "success"
          }
        }
      },
      "BadRequestError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status",
            "example": "error"
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "${error.message}"
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "bad_request"
          }
        }
      },
      "UnauthorizedError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status",
            "example": "error"
          },
          "message": {
            "type": "string",
            "description": "Unauthorized",
            "example": "Usuario o contraseña invalida."
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "authentication_failed"
          }
        }
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status",
            "example": "error"
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Recursos no encontrados."
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "not_found"
          }
        }
      },
      "ConflictError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status",
            "example": "error"
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "El ... ya existe."
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "conflict"
          }
        }
      },
      "InternalServerError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status",
            "example": "error"
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "${error.message}"
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "internal_server_error"
          }
        }
      }
    }
  },
...
}